resove comments

+++ java/src/test/java/ai/rapids/cudf/TableTest.java
@@ -3311,11 +3311,11 @@ public class TableTest extends CudfTestBase {
         .column(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1) // GBY Key
         .column(0, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 2, 2) // GBY Key
         .column(7, 5, 1, 9, 7, 9, 8, 2, 8, 0, 6, 6, 8) // Agg Column
-        .column((long) 1, (long)1, (long)2, (long)3, (long)3, (long)3, (long)4, (long)4, (long)5, (long)5, (long)6, (long)6, (long)7) // orderBy Key
+        .column(1L, 1L, 2L, 3L, 3L, 3L, 4L, 4L, 5L, 5L, 6L, 6L, 7L) // orderBy Key
         .column((short) 1, (short)1, (short)2, (short)3, (short)3, (short)3, (short)4, (short)4, (short)5, (short)5, (short)6, (short)6, (short)7) // orderBy Key
-        .column((int) 1, (int)1, (int)2, (int)3, (int)3, (int)3, (int)4, (int)4, (int)5, (int)5, (int)6, (int)6, (int)7) // orderBy Key
+        .column(1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // orderBy Key
         .column((byte) 1, (byte)1, (byte)2, (byte)3, (byte)3, (byte)3, (byte)4, (byte)4, (byte)5, (byte)5, (byte)6, (byte)6, (byte)7) // orderBy Key
-        .timestampDayColumn( 1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // timestamp orderBy Key
+        .timestampDayColumn(1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // timestamp orderBy Key
         .build()) {
 
       int[] orderByIndices = {3, 4, 5, 6, 7};
@@ -3346,11 +3346,11 @@ public class TableTest extends CudfTestBase {
         .column(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1) // GBY Key
         .column(0, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 2, 2) // GBY Key
         .column(7, 5, 1, 9, 7, 9, 8, 2, 8, 0, 6, 6, 8) // Agg Column
-        .column((long) 1, (long)1, (long)2, (long)3, (long)3, (long)3, (long)4, (long)4, (long)5, (long)5, (long)6, (long)6, (long)7) // orderBy Key
+        .column(1L, 1L, 2L, 3L, 3L, 3L, 4L, 4L, 5L, 5L, 6L, 6L, 7L) // orderBy Key
         .column((short) 1, (short)1, (short)2, (short)3, (short)3, (short)3, (short)4, (short)4, (short)5, (short)5, (short)6, (short)6, (short)7) // orderBy Key
-        .column((int) 1, (int)1, (int)2, (int)3, (int)3, (int)3, (int)4, (int)4, (int)5, (int)5, (int)6, (int)6, (int)7) // orderBy Key
+        .column(1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // orderBy Key
         .column((byte) 1, (byte)1, (byte)2, (byte)3, (byte)3, (byte)3, (byte)4, (byte)4, (byte)5, (byte)5, (byte)6, (byte)6, (byte)7) // orderBy Key
-        .timestampDayColumn( 1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // Timestamp orderBy Key
+        .timestampDayColumn(1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // Timestamp orderBy Key
         .build()) {
 
       int[] orderByIndices = {3, 4, 5, 6, 7};
@@ -3384,11 +3384,11 @@ public class TableTest extends CudfTestBase {
         .column(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1) // GBY Key
         .column(0, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 2, 2) // GBY Key
         .column(7, 5, 1, 9, 7, 9, 8, 2, 8, 0, 6, 6, 8) // Agg Column
-        .column((long) 1, (long)1, (long)2, (long)3, (long)3, (long)3, (long)4, (long)4, (long)5, (long)5, (long)6, (long)6, (long)7) // orderBy Key
+        .column(1L, 1L, 2L, 3L, 3L, 3L, 4L, 4L, 5L, 5L, 6L, 6L, 7L) // orderBy Key
         .column((short) 1, (short)1, (short)2, (short)3, (short)3, (short)3, (short)4, (short)4, (short)5, (short)5, (short)6, (short)6, (short)7) // orderBy Key
-        .column((int) 1, (int)1, (int)2, (int)3, (int)3, (int)3, (int)4, (int)4, (int)5, (int)5, (int)6, (int)6, (int)7) // orderBy Key
+        .column(1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // orderBy Key
         .column((byte) 1, (byte)1, (byte)2, (byte)3, (byte)3, (byte)3, (byte)4, (byte)4, (byte)5, (byte)5, (byte)6, (byte)6, (byte)7) // orderBy Key
-        .timestampDayColumn( 1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // Timestamp orderBy Key
+        .timestampDayColumn(1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // Timestamp orderBy Key
         .build()) {
 
       int[] orderByIndices = {3, 4, 5, 6, 7};
@@ -3431,11 +3431,11 @@ public class TableTest extends CudfTestBase {
         .column(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1) // GBY Key
         .column(0, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 2, 2) // GBY Key
         .column(7, 5, 1, 9, 7, 9, 8, 2, 8, 0, 6, 6, 8) // Agg Column
-        .column((long) 1, (long)1, (long)2, (long)3, (long)3, (long)3, (long)4, (long)4, (long)5, (long)5, (long)6, (long)6, (long)7) // orderBy Key
+        .column(1L, 1L, 2L, 3L, 3L, 3L, 4L, 4L, 5L, 5L, 6L, 6L, 7L) // orderBy Key
         .column((short) 1, (short)1, (short)2, (short)3, (short)3, (short)3, (short)4, (short)4, (short)5, (short)5, (short)6, (short)6, (short)7) // orderBy Key
-        .column((int) 1, (int)1, (int)2, (int)3, (int)3, (int)3, (int)4, (int)4, (int)5, (int)5, (int)6, (int)6, (int)7) // orderBy Key
+        .column(1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // orderBy Key
         .column((byte) 1, (byte)1, (byte)2, (byte)3, (byte)3, (byte)3, (byte)4, (byte)4, (byte)5, (byte)5, (byte)6, (byte)6, (byte)7) // orderBy Key
-        .timestampDayColumn( 1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // Timestamp orderBy Key
+        .timestampDayColumn(1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // Timestamp orderBy Key
         .build()) {
 
       int[] orderByIndices = {3, 4, 5, 6, 7};
@@ -3467,11 +3467,11 @@ public class TableTest extends CudfTestBase {
         .column(1, 1, 1, 1,  1, 1, 1, 1,  1, 1, 1, 1, 1) // GBY Key
         .column(0, 0, 0, 0,  1, 1, 1, 1,  2, 2, 2, 2, 2) // GBY Key
         .column(7, 5, 1, 9,  7, 9, 8, 2,  8, 0, 6, 6, 8) // Agg Column
-        .column((long)7, (long)6, (long)6, (long)5,  (long)5, (long)4, (long)4, (long)3,  (long)3, (long)3, (long)2, (long)1, (long)1)
+        .column(7L, 6L, 6L, 5L, 5L, 4L, 4L, 3L, 3L, 3L, 2L, 1L, 1L)
         .column((short)7, (short)6, (short)6, (short)5,  (short)5, (short)4, (short)4, (short)3,  (short)3, (short)3, (short)2, (short)1, (short)1)
-        .column((int)7, (int)6, (int)6, (int)5,  (int)5, (int)4, (int)4, (int)3,  (int)3, (int)3, (int)2, (int)1, (int)1)
+        .column(7, 6, 6, 5,  5, 4, 4, 3,  3, 3, 2, 1, 1)
         .column((byte)7, (byte)6, (byte)6, (byte)5,  (byte)5, (byte)4, (byte)4, (byte)3,  (byte)3, (byte)3, (byte)2, (byte)1, (byte)1)
-        .timestampDayColumn( 7, 6, 6, 5,  5, 4, 4, 3,  3, 3, 2, 1, 1) // Timestamp Key
+        .timestampDayColumn(7, 6, 6, 5,  5, 4, 4, 3,  3, 3, 2, 1, 1) // Timestamp Key
         .build()) {
       int[] orderByIndices = {3, 4, 5, 6, 7};
       for (int orderIndex : orderByIndices) {
@@ -3512,11 +3512,11 @@ public class TableTest extends CudfTestBase {
   void testRangeWindowingWithoutGroupByColumns() {
     try (Table unsorted = new Table.TestBuilder()
         .column(             7, 5, 1, 9, 7, 9, 8, 2, 8, 0, 6, 6, 8) // Agg Column
-        .column((long) 1, (long)1, (long)2, (long)3, (long)3, (long)3, (long)4, (long)4, (long)5, (long)5, (long)6, (long)6, (long)7) // orderBy Key
+        .column(1L, 1L, 2L, 3L, 3L, 3L, 4L, 4L, 5L, 5L, 6L, 6L, 7L) // orderBy Key
         .column((short) 1, (short)1, (short)2, (short)3, (short)3, (short)3, (short)4, (short)4, (short)5, (short)5, (short)6, (short)6, (short)7) // orderBy Key
-        .column((int) 1, (int)1, (int)2, (int)3, (int)3, (int)3, (int)4, (int)4, (int)5, (int)5, (int)6, (int)6, (int)7) // orderBy Key
+        .column(1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // orderBy Key
         .column((byte) 1, (byte)1, (byte)2, (byte)3, (byte)3, (byte)3, (byte)4, (byte)4, (byte)5, (byte)5, (byte)6, (byte)6, (byte)7) // orderBy Key
-        .timestampDayColumn( 1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // Timestamp orderBy Key
+        .timestampDayColumn(1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // Timestamp orderBy Key
         .build()) {
 
       int[] orderByIndices = {1, 2, 3, 4, 5};
@@ -3587,11 +3587,11 @@ public class TableTest extends CudfTestBase {
         .column(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1) // GBY Key
         .column(0, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 2, 2) // GBY Key
         .column(7, 5, 1, 9, 7, 9, 8, 2, 8, 0, 6, 6, 8) // Agg Column
-        .column((long) 1, (long)1, (long)2, (long)3, (long)3, (long)3, (long)4, (long)4, (long)5, (long)5, (long)6, (long)6, (long)7) // orderBy Key
+        .column(1L, 1L, 2L, 3L, 3L, 3L, 4L, 4L, 5L, 5L, 6L, 6L, 7L) // orderBy Key
         .column((short) 1, (short)1, (short)2, (short)3, (short)3, (short)3, (short)4, (short)4, (short)5, (short)5, (short)6, (short)6, (short)7) // orderBy Key
-        .column((int) 1, (int)1, (int)2, (int)3, (int)3, (int)3, (int)4, (int)4, (int)5, (int)5, (int)6, (int)6, (int)7) // orderBy Key
+        .column(1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // orderBy Key
         .column((byte) 1, (byte)1, (byte)2, (byte)3, (byte)3, (byte)3, (byte)4, (byte)4, (byte)5, (byte)5, (byte)6, (byte)6, (byte)7) // orderBy Key
-        .timestampDayColumn( 1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // Timestamp orderBy Key
+        .timestampDayColumn(1, 1, 2, 3, 3, 3, 4, 4, 5, 5, 6, 6, 7) // Timestamp orderBy Key
         .build()) {
       int[] orderByIndices = {3, 4, 5, 6, 7};
       for (int orderIndex : orderByIndices) {
@@ -3624,7 +3624,7 @@ public class TableTest extends CudfTestBase {
         .column(0, 0, 0, 0, 0, 0,  1, 1, 1, 1, 1, 1, 1) // GBY Key
         .column(7, 5, 1, 9, 7, 9,  8, 2, 8, 0, 6, 6, 8) // Agg Column
         .column( null, null, null, 2, 3, 5,  null, null, 1, 2, 4, 5, 7) // Timestamp Key
-        .column( null, null, null, (long)2, (long)3, (long)5,  null, null, (long)1, (long)2, (long)4, (long)5, (long)7) // orderBy Key
+        .column( null, null, null, 2L, 3L, 5L,  null, null, 1L, 2L, 4L, 5L, 7L) // orderBy Key
         .column( null, null, null, (short)2, (short)3, (short)5,  null, null, (short)1, (short)2, (short)4, (short)5, (short)7) // orderBy Key
         .column( null, null, null, (byte)2, (byte)3, (byte)5,  null, null, (byte)1, (byte)2, (byte)4, (byte)5, (byte)7) // orderBy Key
         .timestampDayColumn( null, null, null, 2, 3, 5,  null, null, 1, 2, 4, 5, 7) // Timestamp orderBy Key
@@ -3703,7 +3703,7 @@ public class TableTest extends CudfTestBase {
         .column(0, 0, 0, 0, 0, 0,  1, 1, 1, 1, 1, 1, 1) // GBY Key
         .column(7, 5, 1, 9, 7, 9,  8, 2, 8, 0, 6, 6, 8) // Agg Column
         .column(null, null, null, 5, 3, 2,  null, null, 7, 5, 4, 2, 1) // Timestamp Key
-        .column(null, null, null, (long)5, (long)3, (long)2,  null, null, (long)7, (long)5, (long)4, (long)2, (long)1) // orderby Key
+        .column(null, null, null, 5L, 3L, 2L,  null, null, 7L, 5L, 4L, 2L, 1L) // orderby Key
         .column(null, null, null, (short)5, (short)3, (short)2,  null, null, (short)7, (short)5, (short)4, (short)2, (short)1) // orderby Key
         .column(null, null, null, (byte)5, (byte)3, (byte)2,  null, null, (byte)7, (byte)5, (byte)4, (byte)2, (byte)1) // orderby Key
         .timestampDayColumn(null, null, null, 5, 3, 2, null, null, 7, 5, 4, 2, 1) // Timestamp orderby Key
@@ -3787,7 +3787,7 @@ public class TableTest extends CudfTestBase {
         .column(0, 0, 0, 0, 0, 0,  1, 1, 1, 1, 1, 1, 1) // GBY Key
         .column(7, 5, 1, 9, 7, 9,  8, 2, 8, 0, 6, 6, 8) // Agg Column
         .column(2, 3, 5, null, null, null,  1, 2, 4, 5, 7, null, null) // Timestamp Key
-        .column((long)2, (long)3, (long)5, null, null, null, (long)1, (long)2, (long)4, (long)5, (long)7, null, null) // order by Key
+        .column(2L, 3L, 5L, null, null, null, 1L, 2L, 4L, 5L, 7L, null, null) // order by Key
         .column((short)2, (short)3, (short)5, null, null, null, (short)1, (short)2, (short)4, (short)5, (short)7, null, null) // order by Key
         .column((byte)2, (byte)3, (byte)5, null, null, null, (byte)1, (byte)2, (byte)4, (byte)5, (byte)7, null, null) // order by Key
         .timestampDayColumn( 2, 3, 5, null, null, null,  1, 2, 4, 5, 7, null, null) // Timestamp order by Key
@@ -3865,7 +3865,7 @@ public class TableTest extends CudfTestBase {
         .column(0, 0, 0, 0, 0, 0,  1, 1, 1, 1, 1, 1, 1) // GBY Key
         .column(7, 5, 1, 9, 7, 9,  8, 2, 8, 0, 6, 6, 8) // Agg Column
         .column( 5, 3, 2, null, null, null, 7, 5, 4, 2, 1, null, null) // Timestamp Key
-        .column((long)5, (long)3, (long)2, null, null, null, (long)7, (long)5, (long)4, (long)2, (long)1, null, null) // Timestamp Key
+        .column(5L, 3L, 2L, null, null, null, 7L, 5L, 4L, 2L, 1L, null, null) // Timestamp Key
         .column((short)5, (short)3, (short)2, null, null, null, (short)7, (short)5, (short)4, (short)2, (short)1, null, null) // Timestamp Key
         .column((byte)5, (byte)3, (byte)2, null, null, null, (byte)7, (byte)5, (byte)4, (byte)2, (byte)1, null, null) // Timestamp Key
         .timestampDayColumn( 5, 3, 2, X, X, X,  7, 5, 4, 2, 1, X, X) // Timestamp Key
